shop_web/requirements.txt
flask
pymysql
pillow
numpy
opencv-python
pyzbar
gunicorn